+++ domain/src/main/kotlin/com/sooum/domain/usecase/UpdateUserFavoriteUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class UpdateUserFavoriteUseCase @Inject constructor(
     private val repository: UserRepository
 ) {
-    suspend operator fun invoke(id: Long, favorite: Boolean) {
+    suspend operator fun invoke(id: Int, favorite: Boolean) {
         return repository.updateUserFavorite(id, favorite)
     }
 }
